src/test/java/dev/dobicinaitis/advent/Day6Test.java
package dev.dobicinaitis.advent;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day6Test {

    private final Puzzle puzzle = new Day6();

    @Test
    void firstPuzzleExampleTest() {
        assertEquals(288L, puzzle.getFirstPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void firstPuzzleMyInputTest() {
        assertEquals(1108800L, puzzle.getFirstPuzzleSolution(puzzle.myInputFilename));
    }

    @Test
    void secondPuzzleExampleTest() {
        assertEquals(71503L, puzzle.getSecondPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void secondPuzzleMyInputTest() {
        assertEquals(36919753L, puzzle.getSecondPuzzleSolution(puzzle.myInputFilename));
    }
}